refactor(test): IntegrationTestSupport 패키지 경로 변경
- `IntegrationTestSupport` 클래스를 `support` 패키지로 이동
- 관련 테스트 파일에서 경로를 업데이트하여 일관성 유지
